carrito y autentificacion funcionando

+++ Server/apps/product/views.py
@@ -44,6 +44,8 @@ class GetProductByName(APIView):
 #crear producto
 class CreateProduct(APIView):
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({'message': 'You must be logged in to add items to the cart'})
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -53,6 +55,8 @@ class CreateProduct(APIView):
 # crear una categoria
 class CreateCategory(APIView):
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({'message': 'You must be logged in to add items to the cart'})
         serializer = CategorySerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -64,6 +68,8 @@ class CreateCategory(APIView):
 # borrar un producto
 class DeleteProduct(APIView):
     def delete(self, request, product_id):
+        if not request.user.is_authenticated:
+            return Response({'message': 'You must be logged in to add items to the cart'})
         try:
             product = Product.objects.get(id=product_id)
             product.delete()
@@ -74,6 +80,8 @@ class DeleteProduct(APIView):
 # borrar una categoria
 class DeleteCategory(APIView):
     def delete(self, request, category_id):
+        if not request.user.is_authenticated:
+            return Response({'message': 'You must be logged in to add items to the cart'})
         try:
             category = Category.objects.get(id=category_id)
             category.delete()
